fix: resolve database module conflict by integrating GitHub Connection model into existing database.py
The PR was creating a new database/ directory that shadowed the existing
database.py file, causing import errors across the entire codebase.

Changes:
- Moved UserGitHubConnection model into database.py
- Moved all GitHub Connection CRUD functions into database.py
- Removed the conflicting database/ directory
- Fixed imports in test_github_connection_model.py (removed hardcoded path)
- Fixed imports in alembic/env.py
- Added missing Optional import for type hints
- Removed unused ForeignKeyConstraint import
- Fixed all ruff linting errors

This resolves the root cause of the CI failures where all backend tests
were failing due to ImportError when trying to import from database module.
