--- resume-api/database/__init__.py
@@ -1,23 +1,0 @@
-"""
-Database module for ResumeAI.
-
-Contains SQLAlchemy models and CRUD operations.
-"""
-
-from .models import UserGitHubConnection
-from .crud import (
-    get_user_github_connection,
-    create_user_github_connection,
-    update_user_github_connection,
-    delete_user_github_connection,
-    get_github_connection_by_github_user_id,
-)
-
-__all__ = [
-    "UserGitHubConnection",
-    "get_user_github_connection",
-    "create_user_github_connection",
-    "update_user_github_connection",
-    "delete_user_github_connection",
-    "get_github_connection_by_github_user_id",
-]
